refactor(tools): 将_record_read_meta方法移至基类 (#57) - 重构优化: 消除代码重复并统一文件读取元数据记录逻辑   * 从`read_tool.py`中移除重复的`_record_read_meta`方法实现   * 从`read_lines_tool.py`中移除重复的`_record_read_meta`方法实现   * 在`base_tool.py`中添加`_record_read_meta`私有方法，调用`FileTracker.get_file_meta`获取元数据   * 新实现通过`self.workspace.db.record_file_read`记录文件的相对路径、修改时间、大小及校验和
