+++ src/workspace/tools/read_tool.py
@@ -1,6 +1,5 @@
 from pathlib import Path
 
-from src.core.file_tracker import FileTracker
 from src.models.tool_error_response import ToolErrorResponse
 from src.workspace.path_validator import PathNotFoundError, WorkspaceBoundaryError
 from src.workspace.tools.base_tool import BaseTool
@@ -58,12 +57,3 @@ class ReadTool(BaseTool):
             return ToolErrorResponse(self.__class__.__name__, err3).to_str()
         except Exception as err:
             return ToolErrorResponse(self.__class__.__name__, err).to_str()
-
-    def _record_read_meta(self, resolved_path: Path) -> None:
-        try:
-            meta = FileTracker.get_file_meta(resolved_path)
-            if meta:
-                rel_path = str(resolved_path.relative_to(self.workspace.root_path))
-                self.workspace.db.record_file_read(rel_path, meta["mtime"], meta["size"], meta["checksum"])
-        except Exception:
-            pass
